+++ cms/trunk/admin/listusers.php
@@ -42,9 +42,12 @@ if (isset($_GET["message"])) {
 
 	$userlist = UserOperations::LoadUsers();
 
-	#if ($result && $result->RowCount() > 0) {
-	if ($userlist && count($userlist) > 0)
-	{
+	$page = 1;
+	if (isset($_GET['page']))$page = $_GET['page'];
+	$limit = 20;
+	echo "<div align=\"right\" class=\"clearbox\">".pagination($page, count($userlist), $limit)."</div>";
+
+	if ($userlist && count($userlist) > 0){
 		echo '<table cellspacing="0" class="admintable">'."\n";
 		echo "<tr>\n";
 		echo "<td>".lang('username')."</td>\n";
@@ -59,22 +62,23 @@ if (isset($_GET["message"])) {
 		$image_true ="<img src=\"../images/cms/true.gif\" alt=\"".lang('true')."\" title=\"".lang('true')."\" border=\"0\">";
 		$image_false ="<img src=\"../images/cms/false.gif\" alt=\"".lang('false')."\" title=\"".lang('false')."\" border=\"0\">";
 
-		#while($row = $result->FetchRow()) {
-		foreach ($userlist as $oneuser)
-		{
-			echo "<tr class=\"$currow\">\n";
-			echo "<td><a href=\"edituser.php?user_id=".$oneuser->id."\">".$oneuser->username."</a></td>\n";
-			echo "<td align=\"center\">".($oneuser->active == 1?$image_true:$image_false)."</td>\n";
-			if ($edit || $userid == $oneuser->id)
-				echo "<td width=\"16\"><a href=\"edituser.php?user_id=".$oneuser->id."\"><img src=\"../images/cms/edit.gif\" width=\"16\" height=\"16\" border=\"0\" alt=\"".lang('edit')."\"></a></td>\n";
-			else
-				echo "<td>&nbsp;</td>\n";
-			if ($remove)
-				echo "<td width=\"16\"><a href=\"deleteuser.php?user_id=".$oneuser->id."\" onclick=\"return confirm('".lang('deleteconfirm')."');\"><img src=\"../images/cms/delete.gif\" width=\"16\" height=\"16\" border=\"0\" alt=\"".lang('delete')."\"></a></td>\n";
-			echo "</tr>\n";
+		$counter=0;
+		foreach ($userlist as $oneuser){
+			if ($counter < $page*$limit && $counter >= ($page*$limit)-$limit) {
+				echo "<tr class=\"$currow\">\n";
+				echo "<td><a href=\"edituser.php?user_id=".$oneuser->id."\">".$oneuser->username."</a></td>\n";
+				echo "<td align=\"center\">".($oneuser->active == 1?$image_true:$image_false)."</td>\n";
+				if ($edit || $userid == $oneuser->id)
+					echo "<td width=\"16\"><a href=\"edituser.php?user_id=".$oneuser->id."\"><img src=\"../images/cms/edit.gif\" width=\"16\" height=\"16\" border=\"0\" alt=\"".lang('edit')."\"></a></td>\n";
+				else
+					echo "<td>&nbsp;</td>\n";
+				if ($remove)
+					echo "<td width=\"16\"><a href=\"deleteuser.php?user_id=".$oneuser->id."\" onclick=\"return confirm('".lang('deleteconfirm')."');\"><img src=\"../images/cms/delete.gif\" width=\"16\" height=\"16\" border=\"0\" alt=\"".lang('delete')."\"></a></td>\n";
+				echo "</tr>\n";
 
-			($currow=="row1"?$currow="row2":$currow="row1");
-
+				($currow=="row1"?$currow="row2":$currow="row1");
+			}
+			$counter++;
 		}
 
 		echo "</table>\n";
